Update raspberrypi-monitoring.yml
Found that the "machine-id" in the zcube/cadvisor container was failing to read the file from the pi.. I added "- /etc/machine-id:/etc/machine-id:ro" (Line 23) in my yml file and the container no longer gives the error when spinning up the container. Forgive me if this is not the most appropriate way to do this as I am a novice at docker, but I am trying to do my part to help out with this fantastic project!